requirements.txt
requests
paho-mqtt
pytz